Revert "disp: msm: sde: Fix 32-bit compilation issues"
This reverts commit 362bd1d923894f03e5696c85bb39b75bcc4e4210.

+++ drivers/media/platform/msm/sde/rotator/sde_rotator_smmu.c
@@ -705,7 +705,7 @@ int sde_smmu_probe(struct platform_device *pdev)
 				sizeof(*dev->dma_parms), GFP_KERNEL);
 
 	dma_set_max_seg_size(dev, DMA_BIT_MASK(32));
-	dma_set_seg_boundary(dev, (unsigned long)DMA_BIT_MASK(64));
+	dma_set_seg_boundary(dev, DMA_BIT_MASK(64));
 
 	iommu_set_fault_handler(sde_smmu->mmu_mapping->domain,
 			sde_smmu_fault_handler, (void *)sde_smmu);
